api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
plaid-python>=9.0.0
alpaca-trade-api>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
